requirements-testing.txt
mccabe==0.5.2
mock==2.0.0
pep8==1.7.0
pep257==0.7.0
pyaml
pylama==7.0.9
pluggy==0.3.1
pytest==3.0.2
pytest-flake8==0.6
pyflakes==1.2.3
flake8==3.0.0
requests>=2.7
pytest-cov==2.3.1
